add desensitize flag -1
