+++ backend/src/main/java/com/hufudb/onedb/backend/entity/request/DFlagRequest.java
@@ -1,0 +1,5 @@
+package com.hufudb.onedb.backend.entity.request;
+
+public class DFlagRequest {
+    public Boolean desensitizeFlag;
+}
