+++ desensitize/pom.xml
@@ -1,0 +1,41 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project
+        xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
+        xmlns="http://maven.apache.org/POM/4.0.0"
+        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <artifactId>onedb</artifactId>
+        <groupId>com.hufudb.onedb</groupId>
+        <version>1.0-SNAPSHOT</version>
+    </parent>
+    <groupId>com.hufudb.onedb</groupId>
+    <artifactId>desensitize</artifactId>
+    <version>1.0-SNAPSHOT</version>
+    <name>desensitize</name>
+    <properties>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+        <maven.compiler.source>11</maven.compiler.source>
+        <maven.compiler.target>11</maven.compiler.target>
+        <grpc.version>1.43.2</grpc.version>
+        <truelicense.version>1.33</truelicense.version>
+    </properties>
+    <dependencies>
+        <dependency>
+            <groupId>com.hufudb.onedb</groupId>
+            <artifactId>proto</artifactId>
+            <version>1.0-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>com.hufudb.onedb</groupId>
+            <artifactId>data</artifactId>
+            <version>1.0-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>com.hufudb.onedb</groupId>
+            <artifactId>plan</artifactId>
+            <version>1.0-SNAPSHOT</version>
+            <scope>compile</scope>
+        </dependency>
+    </dependencies>
+</project>
