+++ owner/pom.xml
@@ -27,6 +27,11 @@
     </dependency>
     <dependency>
       <groupId>com.hufudb.onedb</groupId>
+      <artifactId>desensitize</artifactId>
+      <version>1.0-SNAPSHOT</version>
+    </dependency>
+    <dependency>
+      <groupId>com.hufudb.onedb</groupId>
       <artifactId>plan</artifactId>
       <version>1.0-SNAPSHOT</version>
     </dependency>
